pyproject.toml
[project]
name = "simple-llm-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Aravind", email = "Aravind@ara-research.ai"}
]
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.11",
    "httpx>=0.28.1",
    "openai>=1.65.4",
    "uvicorn>=0.34.0",
    "anthropic>=0.49.0",
    "google-generativeai>=0.8.4",
    "streamlit>=1.43.0",
]

[dependency-groups]
client = [
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.0",
]

dev = [
    "black>=25.1.0",
    "mypy>=1.15.0",
    "ruff>=0.9.10",
]



[tool.ruff]
select = ["E", "F", "I"]
line-length = 88

[tool.black]
line-length = 88

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
